requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
gitpython==3.1.41
pygithub==2.1.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0
python-multipart==0.0.6
